+++ CHANGELOG.md
@@ -6,6 +6,7 @@ title: Changelog
 
 ### Features
 
+- Added a `@reexport` modifier tag to have TypeDoc convert variable/type references as a re-export instead of a new symbol.
 - Introduced `generateOutputsBegin` and `generateOutputsEnd` events on `Application` for plugin use.
 
 ## v0.28.19 (2026-04-12)